refactor: xstate 타입 코드 분리
