+++ src/constants/pages.ts
@@ -1,6 +1,8 @@
-export enum PAGES {
-  CARD_LIST = 'card-list',
-  ADD_CARD = 'add-card',
-  ADD_CARD_SUCCESS = 'add-card-success',
-  EDIT_CARD_NAME = 'edit-card-name',
-}
+export const PAGES = {
+  CARD_LIST: 'card-list',
+  ADD_CARD: 'add-card',
+  ADD_CARD_SUCCESS: 'add-card-success',
+  EDIT_CARD_NAME: 'edit-card-name',
+} as const;
+
+export type PagesType = (typeof PAGES)[keyof typeof PAGES];
