+++ src/machine/cardMachine.ts
@@ -1,11 +1,10 @@
 import { setup, assign } from 'xstate';
-import { CardInfo } from '../types';
+import { CardInfo, CardMachineContext, CardMachineEvent } from '../types';
 import {
   setLocalStorageItem,
   CARD_STORAGE_KEY,
   getLocalStorageItem,
 } from '../utils/localStorage';
-import { PAGES } from '../constants/pages';
 
 const initialCardState: CardInfo = {
   id: '',
@@ -21,34 +20,8 @@ const initialCardState: CardInfo = {
 
 export const cardMachine = setup({
   types: {} as {
-    context: {
-      cardState: CardInfo;
-      cardList: CardInfo[];
-    };
-    events:
-      | { type: 'UPDATE_CARD_NUMBER'; value: CardInfo['numbers'] }
-      | {
-          type: 'UPDATE_EXPIRATION_DATE';
-          value: CardInfo['expiration'];
-        }
-      | { type: 'UPDATE_OWNER'; value: CardInfo['owner'] }
-      | {
-          type: 'UPDATE_SECURITY_CODE';
-          value: CardInfo['securityCode'];
-        }
-      | { type: 'UPDATE_PASSWORD'; value: CardInfo['password'] }
-      | { type: 'UPDATE_NICKNAME'; value: CardInfo['nickname'] }
-      | { type: 'UPDATE_BRAND'; value: CardInfo['brand'] }
-      | { type: 'RESET_CARD_INFO' }
-      | { type: 'SUBMIT_CARD' }
-      | { type: 'SAVE_CARD_LIST'; value: CardInfo }
-      | { type: 'DELETE_CARD'; value: string }
-      | { type: 'GET_CARD_INFO'; value: string }
-      | { type: 'SAVE_TO_LOCALSTORAGE'; value: CardInfo[] }
-      | { type: PAGES.CARD_LIST }
-      | { type: PAGES.ADD_CARD }
-      | { type: PAGES.ADD_CARD_SUCCESS }
-      | { type: PAGES.EDIT_CARD_NAME };
+    context: CardMachineContext;
+    events: CardMachineEvent;
   },
   actions: {
     saveCard: assign({
